requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF             # pip name: pymupdf
spacy
en-core-web-sm      # spaCy model; install via: python -m spacy download en_core_web_sm
sentence-transformers
rank_bm25
scikit-learn
numpy
pydantic
typing-extensions
sqlalchemy
python-dotenv
passlib[bcrypt]
bcrypt==3.2.2
python-jose
pymysql
email-validator
streamlit